Refactor Input component to include dependencies for navigation and adjust styling
